lance-to-qdrant/pyproject.toml
[project]
name = "lance-to-qdrant"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "lancedb>=0.27.1",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
]
